pyproject.toml
[project]
name = "kpi"
version = "0.1.0"
description = ""
authors = [{ name = "WJQ" }]
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "ffmpeg-python>=0.2.0,<0.3",
    "tqdm>=4.66.2,<5",
    "numpy<2",
    "scenedetect>=0.6.2,<0.7",
    "opencv-python>=4.9.0.80,<5",
    "srt>=3.5.3,<4",
    "orjson>=3.9.15,<4",
    "sentence-transformers>=2.5.1,<3",
    "spacy>=3.7.4,<4",
    "cupy-cuda12x>=13.0.0,<14",
    "python-dotenv>=1.0.1,<2",
    "torchvision>=0.17.2,<0.18",
    "transformers>=4.39.3,<5",
    "librosa>=0.10.1,<0.11",
    "click>=8.1.7,<9",
]
